third_party/blink/web_tests/animations/events/animation-events-create.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
  var ev = document.createEvent('AnimationEvent');
  assert_equals(typeof(ev), 'object');
  assert_equals(ev.animationName, '');
  assert_equals(ev.elapsedTime, 0.0);
}, "Test creation of AnimationEvent");

test(() => {
  ev = document.createEvent('MouseEvent');
  assert_equals(typeof(ev), 'object');
  assert_equals(ev.animationName, undefined);
  assert_equals(ev.elapsedTime, undefined);
}, "Create a MouseEvent and make sure it doesn't have AnimationEvent properties");
</script>
